bin/hydiomatic.hy: Add a --warnings option
With the warnings moved to a separate rule set, we need a way to run our
code against them. The --warnings option does just that. It does not run
any other ruleset, just this.

This fixes #7, though not in the way described there, as the results are
not interleaved, but still.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>