rules/syntaxᵒ: Fix a TypeError in one of the kwapply rules
When looking at a kwapply form, where the applied function starts with a
dot, make sure that we have a symbol to attach that function to. If we
don't, another rule shall take precedence. This is accomplished by
guarding the appropriate part of the form with a typeᵒ.

This change fixes #10.

Reported-by: Zack M. Davis
Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>