hydiomatic.rules: Start adding rules for the Grand Cleanup
This introduces a rather ugly hack: a set of rules will be run after
simplification to clean up temporary renames. The reason we do this, is
because we want to skip applying some rules over and over again: so we
transform the expression so that the rule doesn't recognise it
anymore. But in the end, we want to transform it back. The old functions
that do no cleanup are available under a starred name.

The new rules - not part of the default set - suggest modifications that
follow the Grand Language Cleanup.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>